+++ tests/pyincore/analyses/commercialbuildingrecovery/test_commercialbuildingrecovery.py
@@ -1,14 +1,13 @@
 from pyincore import IncoreClient, RepairService, MappingSet
 from pyincore.analyses.commercialbuildingrecovery.commercialbuildingrecovery import CommercialBuildingRecovery
-from pyincore import Dataset
+import pyincore.globals as pyglobals
 
 
 def run_with_base_class():
     # Connect to IN-CORE service
-    client = IncoreClient()
-    # client.clear_cache()
-    # Joplin
-    buildings = "5dbc8478b9219c06dd242c0d"  # ergo:buildingInventoryVer6 5dbc8478b9219c06dd242c0d
+    client = IncoreClient(pyglobals.INCORE_API_DEV_URL)
+    client.clear_cache()
+    buildings = "5df7d0de425e0b00092d0082"  # ergo:buildingInventoryVer6 5dbc8478b9219c06dd242c0d
 
     # Create commercial recovery instance
     com_recovery = CommercialBuildingRecovery(client)
@@ -19,26 +18,18 @@ def run_with_base_class():
     # Create repair service
     repair_service = RepairService(client)
     mapping_set = MappingSet(repair_service.get_mapping(mapping_id))
-
     com_recovery.set_input_dataset('dfr3_mapping_set', mapping_set)
 
     # input datsets ids
-    # sample_damage_states = "612e4038cf04e0131fb6156c"  # 10 samples 28k buildings - MCS output format
-
-    sample_damage_states = Dataset.from_file("Joplin_bldg_failure_SAM_sample_damage_states.csv",
-                                             "incore:sampleDamageState")
-    com_recovery.set_input_dataset("sample_damage_states", sample_damage_states)
-    mcs_failure = Dataset.from_file("Joplin_bldg_failure_SAM_failure_probability.csv",
-                                    "incore:failureProbability")
-    com_recovery.set_input_dataset("mcs_failure", mcs_failure)
-    # delay_factors = "60fb433cd3c92a78c89d21cc"  # DS_0, etc.
-    delay_factors = Dataset.from_file("Dataset1_REDi_business_new.csv",
-                                      "incore:buildingRecoveryFactors")
-    com_recovery.set_input_dataset("delay_factors", delay_factors)
+    sample_damage_states = "64ee146456b25759cfc599ac"  # 10 samples 28k buildings - MCS output format
+    mcs_failure = '64ee144256b25759cfc599a5'
+    delay_factors = "64ee10e756b25759cfc53243"
 
     # Load input datasets
-    # com_recovery.load_remote_input_dataset("sample_damage_states", sample_damage_states)
-    # com_recovery.load_remote_input_dataset("delay_factors", delay_factors)
+    com_recovery.load_remote_input_dataset("sample_damage_states", sample_damage_states)
+    com_recovery.load_remote_input_dataset("mcs_failure", mcs_failure)
+    com_recovery.load_remote_input_dataset("delay_factors", delay_factors)
+
 
     # Input parameters
     num_samples = 10