+++ pyincore/analyses/nonstructbuildingdamage/nonstructbuildingdamage.py
@@ -5,6 +5,7 @@
 # and is available at https://www.mozilla.org/en-US/MPL/2.0/
 
 import concurrent.futures
+from itertools import repeat
 
 from pyincore import AnalysisUtil, GeoUtil
 from pyincore import BaseAnalysis, HazardService, FragilityService
@@ -31,6 +32,9 @@ class NonStructBuildingDamage(BaseAnalysis):
         """Executes building damage analysis."""
         # Building dataset
         building_set = self.get_input_dataset("buildings").get_inventory_reader()
+
+        # get input hazard
+        hazard, hazard_type, hazard_dataset_id = self.create_hazard_object_from_input_params()
 
         # set Default Fragility key
         fragility_key_as = self.get_parameter("fragility_key_as")
@@ -69,7 +73,10 @@ class NonStructBuildingDamage(BaseAnalysis):
 
         (ds_results, damage_results) = self.building_damage_concurrent_future(self.building_damage_analysis_bulk_input,
                                                                               num_workers,
-                                                                              inventory_args)
+                                                                              inventory_args,
+                                                                              repeat(hazard),
+                                                                              repeat(hazard_type),
+                                                                              repeat(hazard_dataset_id))
 
         self.set_result_csv_data("result", ds_results, name=self.get_parameter("result_name"))
         self.set_result_json_data("damage_result",
@@ -99,11 +106,14 @@ class NonStructBuildingDamage(BaseAnalysis):
 
         return output_ds, output_dmg
 
-    def building_damage_analysis_bulk_input(self, buildings):
+    def building_damage_analysis_bulk_input(self, buildings, hazard, hazard_type, hazard_dataset_id):
         """Run analysis for multiple buildings.
 
         Args:
             buildings (list): Multiple buildings from input inventory set.
+            hazard (obj): Hazard object.
+            hazard_type (str): Hazard type.
+            hazard_dataset_id (str): Hazard dataset id.
 
         Returns:
             dict: An ordered dictionary with building damage values.
@@ -111,8 +121,6 @@ class NonStructBuildingDamage(BaseAnalysis):
 
         """
         # read static parameters from object self
-        hazard_type = self.get_parameter("hazard_type")
-        hazard_dataset_id = self.get_parameter("hazard_id")
         liq_geology_dataset_id = self.get_parameter("liq_geology_dataset_id")
         use_liquefaction = self.get_parameter("use_liquefaction")
         use_hazard_uncertainty = self.get_parameter("use_hazard_uncertainty")
@@ -177,8 +185,8 @@ class NonStructBuildingDamage(BaseAnalysis):
 
         # get hazard values and liquefaction
         if hazard_type == 'earthquake':
-            hazard_resp_as = self.hazardsvc.post_earthquake_hazard_values(hazard_dataset_id, values_payload_as)
-            hazard_resp_ds = self.hazardsvc.post_earthquake_hazard_values(hazard_dataset_id, values_payload_ds)
+            hazard_resp_as = hazard.read_hazard_values(values_payload_as, self.hazardsvc)
+            hazard_resp_ds = hazard.read_hazard_values(values_payload_ds, self.hazardsvc)
 
             # adjust dmg probability for liquefaction
             if use_liquefaction:
@@ -345,13 +353,13 @@ class NonStructBuildingDamage(BaseAnalysis):
                 },
                 {
                     'id': 'hazard_type',
-                    'required': True,
+                    'required': False,
                     'description': 'Hazard Type (e.g. earthquake)',
                     'type': str
                 },
                 {
                     'id': 'hazard_id',
-                    'required': True,
+                    'required': False,
                     'description': 'Hazard ID',
                     'type': str
                 },
@@ -391,6 +399,14 @@ class NonStructBuildingDamage(BaseAnalysis):
                     'required': False,
                     'description': 'If using parallel execution, the number of cpus to request',
                     'type': int
+                },
+            ],
+            'input_hazards': [
+                {
+                    'id': 'hazard',
+                    'required': False,
+                    'description': 'Hazard object',
+                    'type': ["earthquake"]
                 },
             ],
             'input_datasets': [