+++ docs/source/conf.py
@@ -33,9 +33,9 @@ project = 'pyIncore'
 author = ''
 
 # The short X.Y version
-version = '1.12'
+version = '1.13'
 # The full version, including alpha/beta/rc tags
-release = '1.12.0'
+release = '1.13.0'
 
 # -- General configuration ---------------------------------------------------