+++ pyincore/analyses/bridgedamage/bridgedamage.py
@@ -33,9 +33,9 @@ class BridgeDamage(BaseAnalysis):
         # Bridge dataset
         bridge_set = self.get_input_dataset("bridges").get_inventory_reader()
 
-        # Get hazard input
-        hazard_type = self.get_parameter("hazard_type")
-        hazard_dataset_id = self.get_parameter("hazard_id")
+        # get input hazard
+        hazard, hazard_type, hazard_dataset_id = self.create_hazard_object_from_input_params()
+
         user_defined_cpu = 1
 
         if not self.get_parameter("num_cpu") is None and self.get_parameter(
@@ -55,8 +55,7 @@ class BridgeDamage(BaseAnalysis):
             count += avg_bulk_input_size
 
         (ds_results, damage_results) = self.bridge_damage_concurrent_future(
-            self.bridge_damage_analysis_bulk_input, num_workers,
-            inventory_args, repeat(hazard_type),
+            self.bridge_damage_analysis_bulk_input, num_workers, inventory_args, repeat(hazard), repeat(hazard_type),
             repeat(hazard_dataset_id))
 
         self.set_result_csv_data("result", ds_results, name=self.get_parameter("result_name"))
@@ -88,19 +87,20 @@ class BridgeDamage(BaseAnalysis):
 
         return output_ds, output_dmg
 
-    def bridge_damage_analysis_bulk_input(self, bridges, hazard_type,
-                                          hazard_dataset_id):
+    def bridge_damage_analysis_bulk_input(self, bridges, hazard, hazard_type, hazard_dataset_id):
         """Run analysis for multiple bridges.
 
         Args:
             bridges (list): Multiple bridges from input inventory set.
-            hazard_type (str): Hazard type, either earthquake, tornado, tsunami, or hurricane.
-            hazard_dataset_id (str): An id of the hazard exposure.
+            hazard (obj): Hazard object.
+            hazard_type (str): Type of hazard.
+            hazard_dataset_id (str): ID of hazard.
 
         Returns:
             list: A list of ordered dictionaries with bridge damage values and other data/metadata.
 
         """
+
         # Get Fragility key
         fragility_key = self.get_parameter("fragility_key")
         if fragility_key is None:
@@ -161,18 +161,7 @@ class BridgeDamage(BaseAnalysis):
         # not needed anymore as they are already split into mapped and unmapped
         del bridges
 
-        if hazard_type == 'earthquake':
-            hazard_vals = self.hazardsvc.post_earthquake_hazard_values(hazard_dataset_id, values_payload)
-        elif hazard_type == 'tornado':
-            hazard_vals = self.hazardsvc.post_tornado_hazard_values(hazard_dataset_id, values_payload)
-        elif hazard_type == 'tsunami':
-            hazard_vals = self.hazardsvc.post_tsunami_hazard_values(hazard_dataset_id, values_payload)
-        elif hazard_type == 'hurricane':
-            hazard_vals = self.hazardsvc.post_hurricane_hazard_values(hazard_dataset_id, values_payload)
-        elif hazard_type == 'flood':
-            hazard_vals = self.hazardsvc.post_flood_hazard_values(hazard_dataset_id, values_payload)
-        else:
-            raise ValueError("The provided hazard type is not supported yet by this analysis")
+        hazard_vals = hazard.read_hazard_values(values_payload, self.hazardsvc)
 
         # Check if liquefaction is applicable
         if use_liquefaction and geology_dataset_id is not None:
@@ -188,6 +177,7 @@ class BridgeDamage(BaseAnalysis):
             damage_result = dict()
             dmg_probability = dict()
             dmg_intervals = dict()
+            ground_failure_prob = None
             selected_fragility_set = fragility_set[bridge["id"]]
 
             if isinstance(selected_fragility_set.fragility_curves[0], DFR3Curve):
@@ -303,18 +293,6 @@ class BridgeDamage(BaseAnalysis):
                     'type': str
                 },
                 {
-                    'id': 'hazard_type',
-                    'required': True,
-                    'description': 'Hazard Type (e.g. earthquake)',
-                    'type': str
-                },
-                {
-                    'id': 'hazard_id',
-                    'required': True,
-                    'description': 'Hazard ID',
-                    'type': str
-                },
-                {
                     'id': 'fragility_key',
                     'required': False,
                     'description': 'Fragility key to use in mapping dataset',
@@ -343,6 +321,26 @@ class BridgeDamage(BaseAnalysis):
                     'required': False,
                     'description': 'If using parallel execution, the number of cpus to request',
                     'type': int
+                },
+                {
+                    'id': 'hazard_id',
+                    'required': False,
+                    'description': 'Hazard object id',
+                    'type': str
+                },
+                {
+                    'id': 'hazard_type',
+                    'required': False,
+                    'description': 'Hazards type',
+                    'type': str
+                },
+            ],
+            'input_hazards': [
+                {
+                    'id': 'hazard',
+                    'required': False,
+                    'description': 'Hazard object',
+                    'type': ["earthquake", "tornado", "hurricane", "flood", "tsunami"]
                 },
             ],
             'input_datasets': [