+++ pyincore/analyses/pipelinedamage/pipelinedamage.py
@@ -35,8 +35,8 @@ class PipelineDamage(BaseAnalysis):
         pipeline_dataset = self.get_input_dataset("pipeline").get_inventory_reader()
 
         # Get hazard input
-        hazard_type = self.get_parameter("hazard_type")
-        hazard_dataset_id = self.get_parameter("hazard_id")
+        hazard, hazard_type, hazard_dataset_id = self.create_hazard_object_from_input_params()
+
         user_defined_cpu = 1
 
         if not self.get_parameter("num_cpu") is None and self.get_parameter(
@@ -58,7 +58,7 @@ class PipelineDamage(BaseAnalysis):
 
         (results, damage_results) = self.pipeline_damage_concurrent_future(
             self.pipeline_damage_analysis_bulk_input, num_workers,
-            inventory_args, repeat(hazard_type), repeat(hazard_dataset_id))
+            inventory_args, repeat(hazard), repeat(hazard_type), repeat(hazard_dataset_id))
 
         self.set_result_csv_data("result", results, name=self.get_parameter("result_name"))
         self.set_result_json_data("metadata",
@@ -90,12 +90,12 @@ class PipelineDamage(BaseAnalysis):
 
         return output_ds, output_dmg
 
-    def pipeline_damage_analysis_bulk_input(self, pipelines, hazard_type,
-                                            hazard_dataset_id):
+    def pipeline_damage_analysis_bulk_input(self, pipelines, hazard, hazard_type, hazard_dataset_id):
         """Run pipeline damage analysis for multiple pipelines.
 
         Args:
             pipelines (list): Multiple pipelines from pipeline dataset.
+            hazard (obj): Hazard object.
             hazard_type (str): Hazard type (earthquake or tsunami).
             hazard_dataset_id (str): An id of the hazard exposure.
 
@@ -143,16 +143,8 @@ class PipelineDamage(BaseAnalysis):
         # not needed anymore as they are already split into mapped and unmapped
         del pipelines
 
-        if hazard_type == 'earthquake':
-            hazard_vals = self.hazardsvc.post_earthquake_hazard_values(hazard_dataset_id, values_payload)
-        elif hazard_type == 'tornado':
-            raise ValueError("The provided hazard type is not supported yet by this analysis")
-        elif hazard_type == 'tsunami':
-            hazard_vals = self.hazardsvc.post_tsunami_hazard_values(hazard_dataset_id, values_payload)
-        elif hazard_type == 'hurricane':
-            raise ValueError("The provided hazard type is not supported yet by this analysis")
-        elif hazard_type == 'flood':
-            raise ValueError("The provided hazard type is not supported yet by this analysis")
+        if hazard_type == 'earthquake' or "tsunami":
+            hazard_vals = hazard.read_hazard_values(values_payload, self.hazardsvc)
         else:
             raise ValueError("The provided hazard type is not supported yet by this analysis")
 
@@ -239,13 +231,13 @@ class PipelineDamage(BaseAnalysis):
                 },
                 {
                     'id': 'hazard_type',
-                    'required': True,
+                    'required': False,
                     'description': 'Hazard Type',
                     'type': str
                 },
                 {
                     'id': 'hazard_id',
-                    'required': True,
+                    'required': False,
                     'description': 'Hazard ID',
                     'type': str
                 },
@@ -267,6 +259,14 @@ class PipelineDamage(BaseAnalysis):
                     'description': 'Geology dataset id',
                     'type': str,
                 }
+            ],
+            'input_hazards': [
+                {
+                    'id': 'hazard',
+                    'required': False,
+                    'description': 'Hazard object',
+                    'type': ["earthquake", "tsunami"]
+                },
             ],
             'input_datasets': [
                 {