+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - Add commercial recovery analysis [#395](https://github.com/IN-CORE/pyincore/issues/395)
+- Capability to support for local hazard [#404](https://github.com/IN-CORE/pyincore/issues/404)
+- Add support for local hazard with backward compatibility to analyses [#415](https://github.com/IN-CORE/pyincore/issues/415)
+
 
 ### Fixed
 - Aggregate hazard exposure column for non-structural building damage analysis to avoid column name cutoff and chaining issue with mean damage [#393](https://github.com/IN-CORE/pyincore/issues/393)